Move Memory store to a different folder (#3736)
