+++ src/Neo/Persistence/Providers/MemorySnapshot.cs
@@ -1,6 +1,6 @@
 // Copyright (C) 2015-2025 The Neo Project.
 //
-// MemoryStore.cs file belongs to the neo project and is free
+// MemorySnapshot.cs file belongs to the neo project and is free
 // software distributed under the MIT software license, see the
 // accompanying file LICENSE in the main directory of the
 // repository or http://www.opensource.org/licenses/mit-license.php
@@ -14,37 +14,50 @@
 using Neo.Extensions;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
-using System.Runtime.CompilerServices;
 
-namespace Neo.Persistence
+namespace Neo.Persistence.Providers
 {
     /// <summary>
-    /// An in-memory <see cref="IStore"/> implementation that uses ConcurrentDictionary as the underlying storage.
+    /// <remarks>On-chain write operations on a snapshot cannot be concurrent.</remarks>
     /// </summary>
-    public class MemoryStore : IStore
+    internal class MemorySnapshot : IStoreSnapshot
     {
-        private readonly ConcurrentDictionary<byte[], byte[]> _innerData = new(ByteArrayEqualityComparer.Default);
+        private readonly ConcurrentDictionary<byte[], byte[]> _innerData;
+        private readonly ImmutableDictionary<byte[], byte[]> _immutableData;
+        private readonly ConcurrentDictionary<byte[], byte[]?> _writeBatch;
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public IStore Store { get; }
+
+        internal MemorySnapshot(MemoryStore store, ConcurrentDictionary<byte[], byte[]> innerData)
+        {
+            Store = store;
+            _innerData = innerData;
+            _immutableData = innerData.ToImmutableDictionary(ByteArrayEqualityComparer.Default);
+            _writeBatch = new ConcurrentDictionary<byte[], byte[]?>(ByteArrayEqualityComparer.Default);
+        }
+
+        public void Commit()
+        {
+            foreach (var pair in _writeBatch)
+                if (pair.Value is null)
+                    _innerData.TryRemove(pair.Key, out _);
+                else
+                    _innerData[pair.Key] = pair.Value;
+        }
+
         public void Delete(byte[] key)
         {
-            _innerData.TryRemove(key, out _);
+            _writeBatch[key] = null;
         }
 
         public void Dispose() { }
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public IStoreSnapshot GetSnapshot()
-        {
-            return new MemorySnapshot(this, _innerData);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public void Put(byte[] key, byte[] value)
         {
-            _innerData[key[..]] = value[..];
+            _writeBatch[key[..]] = value[..];
         }
 
         /// <inheritdoc/>
@@ -54,7 +67,7 @@ namespace Neo.Persistence
             if (direction == SeekDirection.Backward && keyOrPrefix.Length == 0) yield break;
 
             var comparer = direction == SeekDirection.Forward ? ByteArrayComparer.Default : ByteArrayComparer.Reverse;
-            IEnumerable<KeyValuePair<byte[], byte[]>> records = _innerData;
+            IEnumerable<KeyValuePair<byte[], byte[]>> records = _immutableData;
             if (keyOrPrefix.Length > 0)
                 records = records
                     .Where(p => comparer.Compare(p.Key, keyOrPrefix) >= 0);
@@ -63,29 +76,20 @@ namespace Neo.Persistence
                 yield return (pair.Key[..], pair.Value[..]);
         }
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public byte[]? TryGet(byte[] key)
         {
-            if (!_innerData.TryGetValue(key, out var value)) return null;
-            return value[..];
+            _immutableData.TryGetValue(key, out var value);
+            return value?[..];
         }
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool TryGet(byte[] key, [NotNullWhen(true)] out byte[]? value)
         {
-            return _innerData.TryGetValue(key, out value);
+            return _immutableData.TryGetValue(key, out value);
         }
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool Contains(byte[] key)
         {
-            return _innerData.ContainsKey(key);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal void Reset()
-        {
-            _innerData.Clear();
+            return _immutableData.ContainsKey(key);
         }
     }
 }
