+++ src/Neo/Persistence/Providers/MemoryStore.cs
@@ -1,6 +1,6 @@
 // Copyright (C) 2015-2025 The Neo Project.
 //
-// MemorySnapshot.cs file belongs to the neo project and is free
+// MemoryStore.cs file belongs to the neo project and is free
 // software distributed under the MIT software license, see the
 // accompanying file LICENSE in the main directory of the
 // repository or http://www.opensource.org/licenses/mit-license.php
@@ -14,50 +14,37 @@
 using Neo.Extensions;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using System.Runtime.CompilerServices;
 
-namespace Neo.Persistence
+namespace Neo.Persistence.Providers
 {
     /// <summary>
-    /// <remarks>On-chain write operations on a snapshot cannot be concurrent.</remarks>
+    /// An in-memory <see cref="IStore"/> implementation that uses ConcurrentDictionary as the underlying storage.
     /// </summary>
-    internal class MemorySnapshot : IStoreSnapshot
+    public class MemoryStore : IStore
     {
-        private readonly ConcurrentDictionary<byte[], byte[]> _innerData;
-        private readonly ImmutableDictionary<byte[], byte[]> _immutableData;
-        private readonly ConcurrentDictionary<byte[], byte[]?> _writeBatch;
+        private readonly ConcurrentDictionary<byte[], byte[]> _innerData = new(ByteArrayEqualityComparer.Default);
 
-        public IStore Store { get; }
-
-        internal MemorySnapshot(MemoryStore store, ConcurrentDictionary<byte[], byte[]> innerData)
-        {
-            Store = store;
-            _innerData = innerData;
-            _immutableData = innerData.ToImmutableDictionary(ByteArrayEqualityComparer.Default);
-            _writeBatch = new ConcurrentDictionary<byte[], byte[]?>(ByteArrayEqualityComparer.Default);
-        }
-
-        public void Commit()
-        {
-            foreach (var pair in _writeBatch)
-                if (pair.Value is null)
-                    _innerData.TryRemove(pair.Key, out _);
-                else
-                    _innerData[pair.Key] = pair.Value;
-        }
-
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public void Delete(byte[] key)
         {
-            _writeBatch[key] = null;
+            _innerData.TryRemove(key, out _);
         }
 
         public void Dispose() { }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public IStoreSnapshot GetSnapshot()
+        {
+            return new MemorySnapshot(this, _innerData);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public void Put(byte[] key, byte[] value)
         {
-            _writeBatch[key[..]] = value[..];
+            _innerData[key[..]] = value[..];
         }
 
         /// <inheritdoc/>
@@ -67,7 +54,7 @@ namespace Neo.Persistence
             if (direction == SeekDirection.Backward && keyOrPrefix.Length == 0) yield break;
 
             var comparer = direction == SeekDirection.Forward ? ByteArrayComparer.Default : ByteArrayComparer.Reverse;
-            IEnumerable<KeyValuePair<byte[], byte[]>> records = _immutableData;
+            IEnumerable<KeyValuePair<byte[], byte[]>> records = _innerData;
             if (keyOrPrefix.Length > 0)
                 records = records
                     .Where(p => comparer.Compare(p.Key, keyOrPrefix) >= 0);
@@ -76,20 +63,29 @@ namespace Neo.Persistence
                 yield return (pair.Key[..], pair.Value[..]);
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public byte[]? TryGet(byte[] key)
         {
-            _immutableData.TryGetValue(key, out var value);
-            return value?[..];
+            if (!_innerData.TryGetValue(key, out var value)) return null;
+            return value[..];
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool TryGet(byte[] key, [NotNullWhen(true)] out byte[]? value)
         {
-            return _immutableData.TryGetValue(key, out value);
+            return _innerData.TryGetValue(key, out value);
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool Contains(byte[] key)
         {
-            return _immutableData.ContainsKey(key);
+            return _innerData.ContainsKey(key);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal void Reset()
+        {
+            _innerData.Clear();
         }
     }
 }
