+++ tests/Neo.Cryptography.MPTTrie.Tests/Cryptography/MPTTrie/UT_Cache.cs
@@ -11,7 +11,7 @@
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo.Extensions;
-using Neo.Persistence;
+using Neo.Persistence.Providers;
 using System.Text;
 
 namespace Neo.Cryptography.MPTTrie.Tests
