+++ tests/Neo.Plugins.RpcServer.Tests/UT_RpcServer.cs
@@ -11,7 +11,7 @@
 
 using Microsoft.AspNetCore.Http;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Neo.Persistence;
+using Neo.Persistence.Providers;
 using Neo.SmartContract;
 using Neo.SmartContract.Native;
 using Neo.UnitTests;
