+++ tests/Neo.Plugins.Storage.Tests/StoreTest.cs
@@ -13,6 +13,7 @@
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo.Persistence;
+using Neo.Persistence.Providers;
 using System.IO;
 using System.Linq;
